recommendation_service/requirements.txt
fastapi
openai
requests
pydantic
python-dotenv
uvicorn